Adds trait for detecting if a type has serializable properties
